+++ data/src/main/java/io/gnosis/data/models/SafeInfo.kt
@@ -6,5 +6,5 @@ import java.math.BigInteger
 data class SafeInfo(
     val address: Solidity.Address,
     val nonce: BigInteger,
-    val threshold: BigInteger
+    val threshold: Int
 )
